api/requirements.txt
alembic==1.13.2
annotated-types==0.7.0
anyio==4.6.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.6.1
durationpy==0.7
email_validator==2.2.0
fastapi==0.115.0
fastapi-cli==0.0.5
google-auth==2.35.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.4
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
kubernetes==31.0.0
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
oauthlib==3.2.2
orjson==3.10.7
packaging==24.1
pluggy==1.5.0
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic-extra-types==2.9.0
pydantic-settings==2.5.2
pydantic_core==2.23.4
Pygments==2.18.0
pytest==8.3.3
pytest-cov==4.1.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.10
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.8.1
rsa==4.9
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.5
typer==0.12.5
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.2.3
uvicorn==0.32.0
watchfiles==0.24.0
websocket-client==1.8.0
websockets==13.1
python-jose[cryptography]==3.5.0
bcrypt==4.1.2